+++ python/Figure10_2b.py
