+++ python/Figure10_4.py
