+++ python/Figure10_5.py
